requirements.txt
Django==4.2
uvicorn==0.27.0.post1
psycopg~=3.1.17
aiohttp~=3.9.3
ipython~=8.21.0
django-redis~=5.4.0
PyYAML~=6.0.1
